requirements.txt

opencv-python
numpy
torch
torchvision
ultralytics
pytesseract
Pillow
pandas
streamlit
# Optional baseline (Linux): openalpr or openalpr-python-bindings
